seguridad/autorizador/requirements.txt
fastapi==0.115.11
pydantic==2.10.6
PyJWT==2.10.1
uvicorn==0.34.0
cryptography==44.0.2